requirements.txt
python-telegram-bot>=20.0
pandas>=1.3.0
nltk>=3.6.0
scikit-learn>=1.0.0
python-dotenv>=0.19.0